Moved generation of logging content to use same approach as async stuff. Moved generation of development content to use same approach as async stuff. Improved error handling around dashboard-rendering exceptions.
